# introduce content phase