--- a/gateway/conf.d/apicast.conf
+++ b/gateway/conf.d/apicast.conf
@@ -59,6 +59,28 @@ location @out_of_band_authrep_action {
   }
 }
 
+location @upstream {
+  internal;
+
+  rewrite_by_lua_block {
+    require('resty.ctx').apply()
+  }
+
+  proxy_pass $proxy_pass;
+
+  proxy_http_version 1.1;
+  proxy_set_header X-Real-IP  $remote_addr;
+  proxy_set_header Host $http_host;
+  proxy_set_header X-3scale-proxy-secret-token $secret_token;
+  proxy_set_header X-3scale-debug "";
+  proxy_set_header Connection "";
+
+  # these are duplicated so when request is redirected here those phases are executed
+  post_action @out_of_band_authrep_action;
+  body_filter_by_lua_block { require('apicast.executor'):body_filter() }
+  header_filter_by_lua_block { require('apicast.executor'):header_filter() }
+}
+
 location / {
   set $cached_key null;
   set $credentials null;
@@ -85,20 +107,13 @@ location / {
     require('apicast.executor'):rewrite()
   }
   access_by_lua_block { require('apicast.executor'):access() }
-  body_filter_by_lua_block { require('apicast.executor'):body_filter() }
-  header_filter_by_lua_block { require('apicast.executor'):header_filter() }
 
   content_by_lua_block { require('apicast.executor'):content() }
 
-  proxy_pass $proxy_pass;
-  proxy_http_version 1.1;
-  proxy_set_header X-Real-IP  $remote_addr;
-  proxy_set_header Host $http_host;
-  proxy_set_header X-3scale-proxy-secret-token $secret_token;
-  proxy_set_header X-3scale-debug "";
-  proxy_set_header Connection "";
-
+  # these are duplicated so those phases are executed when no internal redirect happens
   post_action @out_of_band_authrep_action;
+  body_filter_by_lua_block { require('apicast.executor'):body_filter() }
+  header_filter_by_lua_block { require('apicast.executor'):header_filter() }
 
   include ../apicast.d/location.d/*.conf;
 }