--- a/t/deprecation-warnings.t
+++ b/t/deprecation-warnings.t
@@ -15,17 +15,26 @@ APIcast should emit deprecation warnings when loading code using the old paths.
       "backend_version": 1,
       "proxy": {
         "policy_chain": [
-	  { "name": "policy.echo" },
-	  { "name": "apicast" }
-	]
+          { "name": "policy.echo" },
+          { "name": "apicast" }
+        ],
+        "proxy_rules": [
+          { "pattern": "/", "http_method": "GET", "metric_system_name": "hits", "delta": 2 }
+        ]
       }
     }
   ]
 }
+--- backend
+  location /transactions/authrep.xml {
+    content_by_lua_block {
+      ngx.exit(200)
+    }
+  }
 --- request
-GET /echo
+GET /echo?user_key=foo
 --- response_body
-GET /echo HTTP/1.1
+GET /echo?user_key=foo HTTP/1.1
 --- error_code: 200
 --- no_error_log
 [error]